+++ src/Services/ResourceService.php
@@ -10,6 +10,7 @@ use UnzerSDK\Constants\IdStrings;
 use UnzerSDK\Exceptions\UnzerApiException;
 use UnzerSDK\Resources\Config;
 use UnzerSDK\Resources\PaymentTypes\Applepay;
+use UnzerSDK\Resources\PaymentTypes\Clicktopay;
 use UnzerSDK\Resources\PaymentTypes\Googlepay;
 use UnzerSDK\Resources\PaymentTypes\Klarna;
 use UnzerSDK\Resources\PaymentTypes\PaylaterDirectDebit;
@@ -824,6 +825,9 @@ class ResourceService implements ResourceServiceInterface
                 break;
             case IdStrings::GOOGLE_PAY:
                 $paymentType = new Googlepay();
+                break;
+            case IdStrings::CLICK_TO_PAY:
+                $paymentType = new Clicktopay();
                 break;
             case IdStrings::HIRE_PURCHASE_DIRECT_DEBIT:
             case IdStrings::INSTALLMENT_SECURED: