+++ src/Unzer.php
@@ -53,7 +53,7 @@ class Unzer implements
     public const BASE_URL = 'api.unzer.com';
     public const API_VERSION = 'v1';
     public const SDK_TYPE = 'UnzerPHP';
-    public const SDK_VERSION = '3.6.0';
+    public const SDK_VERSION = '3.7.0';
 
     /** @var string $key */
     private $key;