+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+Click To Pay payment method is added to SDK.
+
+### Added
+
+* Added `\UnzerSDK\Resources\PaymentTypes\ClickToPay` payment method.
+
 ## [3.6.0](https://github.com/unzerdev/php-sdk/compare/3.5.0..3.6.0)
 
 Twint payment method is added to SDK.
@@ -12,6 +20,7 @@ Twint payment method is added to SDK.
 ### Added
 
 * Added `\UnzerSDK\Resources\PaymentTypes\Twint` payment method.
+
 
 ## [3.5.0](https://github.com/unzerdev/php-sdk/compare/3.4.1..3.5.0)