+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED
+## [3.7.0](https://github.com/unzerdev/php-sdk/compare/3.6.0..3.7.0)
 
-Click To Pay payment method is added to SDK.
+Support for Click To Pay payment method is added to SDK.
 
 ### Added